# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: 3GPP Server API
  description: |
    A JSON-to-NGAP/NAS translation server for conformance testing, fuzzing,
    and security assessment of 5G core networks.

    The server manages gNB and UE contexts, translates JSON requests into
    NGAP/NAS messages over SCTP, and returns decoded responses as JSON.

    ## Workflow

    1. **Create a gNB** — dials SCTP to the AMF, performs NG Setup.
    2. **Create a UE** — provisions credentials (K, OPc), computes SUCI.
    3. **Send NGAP messages** — builds NAS inside NGAP, sends over SCTP,
       returns the decoded AMF response.

    ## IE Control

    The JSON schema mirrors NGAP/NAS IEs 1:1. Every optional IE can be
    individually set, overridden, or omitted — enabling both correct
    procedures and deliberate deviations for fuzz testing.

    ## 3GPP References

    - NGAP: TS 38.413
    - NAS 5GS: TS 24.501
    - 5G Security: TS 33.501
  version: "0.2.0"
  license:
    name: BUSL-1.1
    identifier: BUSL-1.1
paths:
...
  /enb:
    post:
      operationId: createENB
      summary: Create eNB
      description: |
        Dials SCTP to the MME on S1-MME (PPID 18), sends an S1 Setup Request, and
        returns the S1 Setup Response or S1 Setup Failure (TS 36.413 §8.7).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateENBRequest"
      responses:
        "201":
          description: eNB created and S1 Setup completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateENBResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /enb/{enb_id}:
    parameters:
      - $ref: "#/components/parameters/enb_id"
    get:
      operationId: getENB
      summary: Get eNB state
      responses:
        "200":
          description: eNB state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ENBStateResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteENB
      summary: Delete eNB
      description: Closes the SCTP connection and removes the eNB.
      responses:
        "204":
          description: eNB deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /enb/{enb_id}/ue/{ue_id}/uplink:
    parameters:
      - $ref: "#/components/parameters/enb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: sendENBUplink
      summary: Send S1-U uplink user data
      description: |
        Encapsulates a crafted inner IP packet (sourced from the selected bearer's
        UE IP) in a plain S1-U G-PDU and sends it to the S-GW/UPF on that bearer's
        uplink TEID. Exactly one of icmp_echo or udp is used. Requires the eNB
        created with enable_gtpu.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ebi: { type: integer, description: "Bearer to send on; 0 or the default EBI uses the default bearer, else an additional PDN's bearer." }
                src: { type: string, description: "Override the inner source IP (source-spoofing tests); defaults to the bearer's UE IP." }
                teid: { type: integer, description: "Override the uplink TEID (invalid-tunnel tests)." }
                icmp_echo:
                  type: object
                  properties:
                    dst: { type: string }
                    id: { type: integer }
                    seq: { type: integer }
                udp:
                  type: object
                  properties:
                    dst: { type: string }
                    dst_port: { type: integer }
                    src_port: { type: integer }
                    payload_hex: { type: string }
      responses:
        "200":
          description: Sent.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /enb/{enb_id}/ue/{ue_id}/downlink/await:
    parameters:
      - $ref: "#/components/parameters/enb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: awaitENBDownlink
      summary: Await S1-U downlink user data
      description: Blocks for a downlink T-PDU on the selected bearer's eNB TEID and returns the decoded inner packet.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ebi: { type: integer, description: "Bearer to await on; defaults to the default bearer." }
                timeout_ms: { type: integer, description: "How long to wait. Default 5000." }
      responses:
        "200":
          description: Decoded downlink packet.
        "404":
          $ref: "#/components/responses/NotFound"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /enb/{enb_id}/gtpu/echo:
    parameters:
      - $ref: "#/components/parameters/enb_id"
    post:
      operationId: sendENBGTPUEcho
      summary: Send a GTP-U Echo Request to the S-GW (S1-U path management)
      description: |
        Sends a GTP-U Echo Request (TS 29.281 §7.2.1) from the eNB's S1-U endpoint
        to sgw_ip and returns the Echo Response. Requires the eNB to have been
        created with enable_gtpu.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sgw_ip]
              properties:
                sgw_ip: { type: string, description: "S-GW S1-U IP to echo." }
                timeout_ms: { type: integer, description: "How long to wait. Default 5000." }
      responses:
        "200": { description: "Echo Response received." }
        "504": { description: "No Echo Response before the timeout." }
  /enb/{enb_id}/gtpu/error-indication/await:
    parameters:
      - $ref: "#/components/parameters/enb_id"
    post:
      operationId: awaitENBErrorIndication
      summary: Wait for a GTP-U Error Indication from the S-GW/UPF
      description: |
        Blocks for a GTP-U Error Indication, which the S-GW/UPF must send when it
        receives a G-PDU on S1-U for a TEID with no context (TS 29.281 §7.3.1).
        Requires the eNB to have been created with enable_gtpu.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                timeout_ms: { type: integer, description: "How long to wait. Default 5000." }
      responses:
        "200": { description: "Error Indication received." }
        "504": { description: "No Error Indication before the timeout." }
...
components:
  parameters:
...
    ue_id:
      name: ue_id
      in: path
      required: true
      schema:
        type: string
    enb_id:
      name: enb_id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: SCTP or NGAP error communicating with the AMF.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    GatewayTimeout:
      description: Timed out waiting for AMF response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CreateENBRequest:
      type: object
      required: [mme_address]
      properties:
        mme_address:
          type: string
          description: "MME S1-MME SCTP address (host:port)."
          example: "10.3.0.2:36412"
        enb_s1_address:
          type: string
          description: Local bind address for SCTP. Defaults to 0.0.0.0.
        mcc:
          type: string
          pattern: "^\\d{3}$"
          example: "001"
        mnc:
          type: string
          pattern: "^\\d{2,3}$"
          example: "01"
        tac:
          type: string
          description: Tracking Area Code (2-octet hex).
          example: "0001"
        enb_id:
          type: integer
          description: Macro eNB ID (20-bit).
          example: 1
        name:
          type: string
        enable_gtpu:
          type: boolean
          description: Bind an S1-U GTP-U endpoint so the eNB terminates the user-plane data path.
        enb_n3_address:
          type: string
          description: IP the S1-U GTP-U endpoint binds (defaults to enb_s1_address).
        raw_s1ap_pdu:
          type: string
          description: >
            Hex S1AP PDU written verbatim in place of a built S1 Setup Request,
            with no encoding or validation — send any PDU, well-formed or not. A
            raw send whose reply times out returns a null response (HTTP 201)
            rather than an error, modelling an MME that drops a malformed PDU.
        wait_for:
          type: array
          items:
            type: string
          description: >
            Downlink message types to block for. Defaults to S1SetupResponse,
            S1SetupFailure, and ErrorIndication.
        timeout_ms:
          type: integer
          description: Wait bound in milliseconds (default 5000, or 2000 for a raw send).
        skip_s1_setup:
          type: boolean
          description: Open the SCTP association without sending an S1 Setup Request.
    CreateENBResponse:
      type: object
      properties:
        enb_id:
          type: string
          description: Store handle used in subsequent /enb/{enb_id} requests.
        response:
          $ref: "#/components/schemas/S1APResponse"
          description: First matching downlink, or null if none arrived in time.
    ENBStateResponse:
      type: object
      properties:
        id:
          type: string
        mcc:
          type: string
        mnc:
          type: string
        tac:
          type: string
        enb_id:
          type: integer
          description: Macro eNB ID.
        name:
          type: string
    S1APResponse:
      type: object
      properties:
        pdu_type:
          type: string
          enum: [initiating_message, successful_outcome, unsuccessful_outcome]
        message_type:
          type: string
          description: "Decoded message name (e.g. S1SetupResponse, S1SetupFailure)."
        raw_hex:
          type: string
          description: Full S1AP PDU as hex.
        s1_setup_response:
          type: object
          description: Present on an S1SetupResponse; served GUMMEIs and MME capacity.
        s1_setup_failure:
          type: object
          description: Present on an S1SetupFailure; cause and optional time-to-wait.
        unknown_ies:
          type: array
          description: ProtocolIEs the received message carries that its message type does not model, in wire order.
          items:
            type: object
            properties:
              id:
                type: integer
              criticality:
                type: string
                description: >-
                  reject, ignore, or notify: how a receiver must act on an IE it does not
                  comprehend (TS 36.413 §10.3.2). A value outside the ENUMERATED range is
                  rendered Criticality(n).
              value_hex:
                type: string
                description: Raw open-type value bytes as hex.
...